chore(emitter): ! move emitter package from `projects/emitter` to `src` folder
