+++ e2e/favicon.ico
