+++ e2e/tsconfig.app.json
