+++ e2e/tslint.json
@@ -1,18 +1,17 @@
 {
-    "extends": "../../tslint.json",
+    "extends": "../tslint.json",
     "rules": {
         "directive-selector": [
             true,
             "attribute",
-            "dispatcher",
+            "app",
             "camelCase"
         ],
         "component-selector": [
             true,
             "element",
-            "dispatcher",
+            "app",
             "kebab-case"
-        ],
-        "no-non-null-assertion": false
+        ]
     }
 }
