--- projects/emitter/karma.conf.js
@@ -1,52 +1,0 @@
-module.exports = (config) => {
-    const configuration = {
-        basePath: '',
-
-        frameworks: ['jasmine', '@angular-devkit/build-angular'],
-
-        plugins: [
-            require('karma-jasmine'),
-            require('karma-chrome-launcher'),
-            require('karma-jasmine-html-reporter'),
-            require('karma-coverage-istanbul-reporter'),
-            require('@angular-devkit/build-angular/plugins/karma')
-        ],
-
-        client: {
-            clearContext: false
-        },
-
-        coverageIstanbulReporter: {
-            dir: require('path').join(__dirname, '../../coverage'),
-            reports: ['html', 'lcovonly'],
-            fixWebpackSourcePaths: true
-        },
-
-        reporters: ['progress', 'kjhtml'],
-
-        port: 9876,
-
-        colors: true,
-
-        logLevel: config.LOG_INFO,
-
-        autoWatch: true,
-
-        browsers: ['Chrome'],
-
-        singleRun: false,
-
-        customLaunchers: {
-            Chrome_travis_ci: {
-                base: 'Chrome',
-                flags: ['--no-sandbox']
-            }
-        }
-    };
-
-    if (process.env.TRAVIS) {
-        configuration.browsers = ['Chrome_travis_ci'];
-    }
-
-    config.set(configuration);
-};
