--- projects/emitter/ng-package.json
@@ -1,8 +1,0 @@
-{
-    "$schema": "../../node_modules/ng-packagr/ng-package.schema.json",
-    "dest": "../../dist/emitter",
-    "deleteDestPath": false,
-    "lib": {
-        "entryFile": "src/public_api.ts"
-    }
-}
