--- projects/emitter/ng-package.prod.json
@@ -1,7 +1,0 @@
-{
-    "$schema": "../../node_modules/ng-packagr/ng-package.schema.json",
-    "dest": "../../dist/emitter",
-    "lib": {
-        "entryFile": "src/public_api.ts"
-    }
-}
