--- projects/emitter/src/test.ts
@@ -1,11 +1,0 @@
-import 'reflect-metadata';
-import 'zone.js/dist/zone';
-import 'zone.js/dist/zone-testing';
-import { getTestBed } from '@angular/core/testing';
-import { BrowserDynamicTestingModule, platformBrowserDynamicTesting } from '@angular/platform-browser-dynamic/testing';
-
-declare const require: any;
-
-getTestBed().initTestEnvironment(BrowserDynamicTestingModule, platformBrowserDynamicTesting());
-const context = require.context('../tests', true, /\.spec\.ts$/);
-context.keys().map(context);
