--- projects/emitter/tsconfig.spec.json
@@ -1,13 +1,0 @@
-{
-    "extends": "../../tsconfig.json",
-    "compilerOptions": {
-        "types": ["jasmine","node"]
-    },
-    "files": [
-        "src/test.ts"
-    ],
-    "include": [
-        "**/*.spec.ts",
-        "**/*.d.ts"
-    ]
-}
