+++ src/ng-package.json
@@ -1,0 +1,8 @@
+{
+    "$schema": "../node_modules/ng-packagr/ng-package.schema.json",
+    "dest": "../dist/emitter",
+    "deleteDestPath": false,
+    "lib": {
+        "entryFile": "public_api.ts"
+    }
+}
