+++ src/ng-package.prod.json
@@ -1,0 +1,7 @@
+{
+    "$schema": "../node_modules/ng-packagr/ng-package.schema.json",
+    "dest": "../dist/emitter",
+    "lib": {
+        "entryFile": "public_api.ts"
+    }
+}
