+++ src/tsconfig.lib.json
@@ -1,5 +1,5 @@
 {
-    "extends": "../../tsconfig.json",
+    "extends": "../tsconfig.json",
     "compilerOptions": {
         "target": "es2015",
         "module": "es2015",
@@ -17,9 +17,5 @@
         "annotateForClosureCompiler": true,
         "flatModuleId": "AUTOGENERATED",
         "flatModuleOutFile": "AUTOGENERATED"
-    },
-    "exclude": [
-        "src/test.ts",
-        "**/*.spec.ts"
-    ]
+    }
 }
